requirements.txt
python-telegram-bot==22.5
stripe
Flask
pymongo
dnspython
yt-dlp